removed the ModuleSelector
* refactored main script to run the bot
