+++ main.py
@@ -1,19 +1,25 @@
-"""
-Script Executions
-"""
 import os
 
 from dotenv import load_dotenv
 
-from interface import ModuleSelector
+from bot import run
+
+
+def get_user_input() -> list[str]:
+    instruction = ""
+    while len(instruction) == 0:
+        instruction = input(" > ")
+
+    return instruction.split(" ")
+
 
 if __name__ == "__main__":
     """
     Token controls access to the Discord bot.
     For safety, it is passed through an environment variable.
     You can set the environment variable in your shell,
-    or using a `.env` file which is loaded with `load_dotenv`.
+    or using a `.env` file as the variable/key `DISCORD_BOT_TOKEN`.
     """
 
     load_dotenv()
-    ModuleSelector(package='examples', token=os.environ["DISCORD_BOT_TOKEN"]).run()
+    run(token=os.environ["DISCORD_BOT_TOKEN"], input_method=get_user_input)
